requirements.txt
discord.py==2.1.1
gTTS==2.3.1
